Changes intended to cut down on GPU and CPU use: elementary support for CARenderer's -nextFrameTime and -updateBounds.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/quartzcore/trunk@35425 72102866-910b-0410-8b05-ffd578937521

+++ Source/CAAnimation.m
@@ -832,7 +832,7 @@ static GSQuartzCoreQuaternion linearInterpolationQuaternion(GSQuartzCoreQuaterni
           const CGFloat * fromComponents = CGColorGetComponents(from);
           const CGFloat * toComponents = CGColorGetComponents(to);
 
-          int numberOfComponents = CGColorGetNumberOfComponents(from);
+          size_t numberOfComponents = CGColorGetNumberOfComponents(from);
           CGFloat valueComponents[numberOfComponents];
           for (int i = 0; i < numberOfComponents; i++)
             {
